+++ test/oauth2/authorize/http/response/sessionstate.test.js
@@ -9,7 +9,7 @@ var crypto = require('crypto');
 describe('oauth2/authorize/http/response/sessionstate', function() {
   
   it('should be annotated', function() {
-    expect(factory['@implements']).to.equal('http://i.authnomicon.org/oauth2/authorization/http/ResponseParameters');
+    expect(factory['@implements']).to.equal('module:oauth2orize.responseParametersFn');
   });
   
   it('should extend response with session state', function(done) {