+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Renamed `http://i.authnomicon.org/oauth2/authorization/http/ResponseType`
 interface to `module:oauth2orize.RequestProcessor`.
+- Renamed `http://i.authnomicon.org/oauth2/authorization/http/RequestParameters`
+interface to `module:oauth2orize.RequestParametersProcessor`.
 
 ## [0.0.4] - 2021-12-15
 ### Added