+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.5] - 2023-10-20
+
+TODO: Review this for accuracy.
+
 ### Changed
 - Renamed `http://i.authnomicon.org/oauth2/AuthorizationCodeService` interface
 to `module:@authnomicon/oauth2.AuthorizationCodeService`.